Trying to fix create game
