+++ src/main/java/ch/uzh/ifi/hase/soprafs21/controller/GameController.java
@@ -17,21 +17,6 @@ public class GameController {
 
     GameController(GameService gameService) { this.gameService = gameService; }
 
-    // TODO #29 test and refine mapping for API-calls for starting a game
-    // TODO get information from lobby directly
-    // pass information and settings from the lobby to create a game
-
-    @PostMapping("/lobbies/{lobbyId}/start")
-    @ResponseStatus(HttpStatus.CREATED)
-    @ResponseBody
-    public GameGetDTO createGame(@PathVariable Long lobbyId) {
-        // copy the input into a game visible for all players threw the repository
-        Game createdGame = gameService.createGame(lobbyId);
-
-        // convert internal representation of game back to API for client
-        return GameDTOMapper.INSTANCE.convertEntityToGameGetDTO(createdGame);
-    }
-
     // TODO #40 test and refine mapping for sending drawing information
     // pass information to the right picture
     @PutMapping("/game/{gameId}/drawing")
